KAFKA-3149: Address review comments
